+++ src/components/atoms/index.js
@@ -1,21 +1,10 @@
-﻿import Button from './Button/Button';
-import Title from './Title/Title';
-import Fieldset from './Fieldset/Fieldset';
-import Input from './Input/Input';
-import Logo from './Logo/Logo';
-import Paragraph from './Paragraph/Paragraph';
-import ButtonIcon from './ButtonIcon/ButtonIcon';
-import Select from './Select/Select';
-import SubTitle from './SubTitle/SubTitle';
-
-export {
-  Button,
-  Title,
-  Fieldset,
-  Input,
-  Logo,
-  Paragraph,
-  ButtonIcon,
-  Select,
-  SubTitle,
-};
+﻿export { default as Button } from './Button/Button';
+export { default as ButtonIcon } from './ButtonIcon/ButtonIcon';
+export { default as Fieldset } from './Fieldset/Fieldset';
+export { default as Input } from './Input/Input';
+export { default as Logo } from './Logo/Logo';
+export { default as Paragraph } from './Paragraph/Paragraph';
+export { default as Select } from './Select/Select';
+export { default as SubTitle } from './SubTitle/SubTitle';
+export { default as Title } from './Title/Title';
+export { default as SignCard } from './SignCard/SignCard';
