+++ src/components/molecules/Basket/Basket.js
@@ -1,0 +1,15 @@
+﻿import React from 'react';
+import Title from 'components/atoms';
+import { Wrapper, Line } from './styled/StyledBasket';
+
+const Basket = () => {
+  return (
+    <Wrapper>
+      <Title>ShopList</Title>
+      <Line />
+      <p>List</p>
+    </Wrapper>
+  );
+};
+
+export default Basket;
