+++ src/components/molecules/CardList/CardList.js
@@ -1,22 +1,14 @@
 ﻿import React from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import ListItem from './ListItem';
-
-const StyledWrapper = styled.ul`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding-left: 0 !important;
-  margin: 6px 0;
-`;
+import { Wrapper } from './styled/StyledCardList';
+import ListItem from '../ListItem/ListItem';
 
 const CardList = ({ content }) => {
   const list = content.map(({ id, name, stock, unit }) => (
     <ListItem key={id} id={id} name={name} stock={stock} unit={unit} />
   ));
 
-  return <StyledWrapper>{list}</StyledWrapper>;
+  return <Wrapper>{list}</Wrapper>;
 };
 
 CardList.propTypes = {
