+++ src/components/molecules/CardList/styled/StyledCardList.js
@@ -1,0 +1,9 @@
+﻿import styled from 'styled-components';
+
+export const Wrapper = styled.ul`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  padding-left: 0 !important;
+  margin: 6px 0;
+`;
