+++ src/views/PantryPage.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import { firestoreConnect } from 'react-redux-firebase';
 import { compose } from 'redux';
 import Card from '../components/organisms/Card';
-import Header from '../components/molecules/Header';
+import Header from '../components/molecules/Header/Header';
 import UserTemplate from '../templates/UserTemplate';
 import {
   fruits,
